python/requirements.txt
pulse-broker
pika
matplotlib
